Implement DAP_CONNECT and DAP_TRANSFER commands
close #5
close #9

+++ jocd/src/br/org/certi/jocd/dapaccess/CmsisDapCore.java
@@ -24,58 +24,57 @@ public class CmsisDapCore {
   private static final String TAG = "CmsisDapCore";
 
   public static enum CommandId {
-    DAP_INFO(0x00),
-    DAP_LED(0x01),
-    DAP_CONNECT(0x02),
-    DAP_DISCONNECT(0x03),
-    DAP_TRANSFER_CONFIGURE(0x04),
-    DAP_TRANSFER(0x05),
-    DAP_TRANSFER_BLOCK(0x06),
-    DAP_TRANSFER_ABORT(0x07),
-    DAP_WRITE_ABORT(0x08),
-    DAP_DELAY(0x09),
-    DAP_RESET_TARGET(0x0a),
-    DAP_SWJ_PINS(0x10),
-    DAP_SWJ_CLOCK(0x11),
-    DAP_SWJ_SEQUENCE(0x12),
-    DAP_SWD_CONFIGURE(0x13),
-    DAP_JTAG_SEQUENCE(0x14),
-    DAP_JTAG_CONFIGURE(0x15),
-    DAP_JTAG_IDCODE(0x16),
-    DAP_VENDOR0(0x80);
+    DAP_INFO((byte) 0x00),
+    DAP_LED((byte) 0x01),
+    DAP_CONNECT((byte) 0x02),
+    DAP_DISCONNECT((byte) 0x03),
+    DAP_TRANSFER_CONFIGURE((byte) 0x04),
+    DAP_TRANSFER((byte) 0x05),
+    DAP_TRANSFER_BLOCK((byte) 0x06),
+    DAP_TRANSFER_ABORT((byte) 0x07),
+    DAP_WRITE_ABORT((byte) 0x08),
+    DAP_DELAY((byte) 0x09),
+    DAP_RESET_TARGET((byte) 0x0a),
+    DAP_SWJ_PINS((byte) 0x10),
+    DAP_SWJ_CLOCK((byte) 0x11),
+    DAP_SWJ_SEQUENCE((byte) 0x12),
+    DAP_SWD_CONFIGURE((byte) 0x13),
+    DAP_JTAG_SEQUENCE((byte) 0x14),
+    DAP_JTAG_CONFIGURE((byte) 0x15),
+    DAP_JTAG_IDCODE((byte) 0x16),
+    DAP_VENDOR0((byte) 0x80);
 
-    public final int value;
+    public final byte value;
 
-    CommandId(int id) {
+    CommandId(byte id) {
       this.value = id;
     }
 
-    public int getValue() {
+    public byte getValue() {
       return value;
     }
   }
 
   public static enum IdInfo {
-    VENDOR_ID(0x01),
-    PRODUCT_ID(0x02),
-    SERIAL_NUMBER(0x03),
-    CMSIS_DAP_FW_VERSION(0x04),
-    TARGET_DEVICE_VENDOR(0x05),
-    TARGET_DEVICE_NAME(0x06),
-    CAPABILITIES(0xf0),
-    SWO_BUFFER_SIZE(0xfd),
-    PACKET_COUNT(0xfe),
-    PACKET_SIZE(0xff);
+    VENDOR_ID((byte) 0x01),
+    PRODUCT_ID((byte) 0x02),
+    SERIAL_NUMBER((byte) 0x03),
+    CMSIS_DAP_FW_VERSION((byte) 0x04),
+    TARGET_DEVICE_VENDOR((byte) 0x05),
+    TARGET_DEVICE_NAME((byte) 0x06),
+    CAPABILITIES((byte) 0xf0),
+    SWO_BUFFER_SIZE((byte) 0xfd),
+    PACKET_COUNT((byte) 0xfe),
+    PACKET_SIZE((byte) 0xff);
 
-    public final int value;
+    public final byte value;
 
-    IdInfo(int id) {
+    IdInfo(byte id) {
       this.value = id;
     }
 
-    public int getValue() {
+    public byte getValue() {
       return value;
     }
   }
-
 }
